refactor(html): move webpack's own embedded renderer out of the printer
The printer ran no CSS through it any more, but still held the code that
did: `builtinEmbeddedRenderer` reached for the CSS `SourceProcessor` from
`lib/html/syntax.js`. It and `stripJsonWhitespace` move to a module of
their own, so the printer file names no minifier but the HTML one, and
`html.syntax` re-exports both where callers already look. Its remaining
use of the CSS module is the lexer `parseCssUrls` reads `url()` with.
