+++ lib/html/syntax.js
@@ -14,7 +14,6 @@ const GenericSourceProcessor = require("../util/SourceProcessor");
 
 const { deferredWrite, declineDeferredWrites } = GenericSourceProcessor;
 
-/** @import { CssProcessOptions } from "../css/syntax" */
 /** @import { OutputHtmlOptions } from "../../declarations/WebpackOptions" */
 
 /**
@@ -74,6 +73,10 @@ const EMBEDDED_LANGUAGES = [
 	HTML_TYPE
 ];
 
+const {
+	builtinEmbeddedRenderer,
+	stripJsonWhitespace
+} = require("./builtinEmbeddedRenderer");
 const {
 	ADDRESS_DIV_P,
 	APPLET_MARQUEE_OBJECT,
@@ -8968,12 +8971,6 @@ const mirrorSelectedContent = (node, select) => {
  * @typedef {{ transforms?: HtmlTransformOptions, collapseWhitespace?: boolean | "conservative" | "smart" | "all", mergeStyles?: boolean, mergeScripts?: boolean, removeEmptyAttributes?: boolean, removeEmptyElements?: boolean, removeRedundantAttributes?: boolean | "smart" | "all", sortAttributes?: boolean, sortTokenLists?: boolean, removeImpliedTags?: boolean | "smart" | "all", renderEmbeddedSource?: EmbeddedSourceRenderer, deferEmbeddedSource?: DeferredEmbeddedSource[], deferSrcdoc?: boolean }} HtmlPrintOptions
  */
 
-/**
- * What `builtinEmbeddedRenderer` minifies inline CSS with: the CSS minifier's
- * own options, so an inline declaration is held to the rules a `.css` asset is.
- * @typedef {Pick<CssProcessOptions, "environment" | "convertLengthUnits" | "rewriteCustomProperties" | "transforms" | "unusedSymbols" | "pseudoClasses">} BuiltinEmbeddedRendererOptions
- */
-
 /** @typedef {import("../util/SourceProcessor").PrintContext<HtmlPath, HtmlNodeRef, HtmlPrintOptions>} PrintContext */
 
 // Whether the text ends inside an unterminated character reference, so the next
@@ -10398,46 +10395,6 @@ const _JSON_SUBTYPE_REGEXP =
  */
 const _minifyInlineJson = (json) =>
 	json.trim() === "" ? json : _renderEmbedded(json, JSON_TYPE);
-
-/**
- * Strip the whitespace between a JSON body's tokens. Every literal is copied
- * byte for byte — re-serializing would reorder nothing but would round numbers
- * through a double, drop a duplicate key and rewrite escapes, none of which is
- * this transform's to do. Escapes are also the security case: unescaping a
- * `<` lets a string close the `<script>` early.
- * @param {string} json a `<script>` body
- * @returns {string} the stripped body, or the body as written where it is not JSON
- */
-const stripJsonWhitespace = (json) => {
-	try {
-		JSON.parse(json);
-	} catch (_err) {
-		// Not JSON after all (a template, a placeholder) — not ours to touch.
-		return json;
-	}
-	let out = "";
-	let inString = false;
-	let escaped = false;
-	for (let i = 0; i < json.length; i++) {
-		const c = json.charCodeAt(i);
-		if (inString) {
-			out += json[i];
-			if (escaped) escaped = false;
-			else if (c === 0x5c) escaped = true;
-			else if (c === 0x22) inString = false;
-			continue;
-		}
-		if (c === 0x22) {
-			inString = true;
-			out += json[i];
-			continue;
-		}
-		// JSON whitespace (RFC 8259): tab, LF, CR, space.
-		if (c === 0x09 || c === 0x0a || c === 0x0d || c === 0x20) continue;
-		out += json[i];
-	}
-	return out;
-};
 
 /**
  * Whether a `<style>` element's body is CSS: the attribute is optional, and the
@@ -13535,49 +13492,10 @@ const parseMsapplicationTask = (input) => {
 };
 
 // CSS syntax is loaded on first `parseCssUrls` call, not at module load —
-// `HtmlGenerator` deliberately defers it the same way, and most documents
-// never carry a URL-bearing SVG presentation attribute.
+// `HtmlGenerator` defers it the same way, and most documents carry no
+// URL-bearing SVG presentation attribute.
 /** @type {typeof import("../css/syntax") | undefined} */
 let _cssSyntax;
-
-/**
- * The renderer webpack ships for what a document embeds: its own CSS minifier
- * for a `<style>` or `style=""` and `stripJsonWhitespace` for JSON, declining
- * every other language. A caller passes it where it has no renderer of its own,
- * or behind one for the languages that one declines.
- * @param {BuiltinEmbeddedRendererOptions=} options the CSS minifier's options, so an inline declaration is minified as the `.css` assets are
- * @returns {EmbeddedSourceRenderer} the renderer
- */
-const builtinEmbeddedRenderer = (options = {}) => {
-	const sheet = { mode: /** @type {"minify"} */ ("minify"), ...options };
-	const block = { ...sheet, as: BLOCK_CONTENTS };
-	// A `style=""` repeats across a document far more often than it varies.
-	/** @type {Map<string, string>} */
-	const blocks = new Map();
-	return (source, info) => {
-		if (info.type === JSON_TYPE) return stripJsonWhitespace(source);
-		if (info.type !== CSS_TYPE) return undefined;
-		const isBlock = info.as === BLOCK_CONTENTS;
-		if (isBlock) {
-			const memoized = blocks.get(source);
-			if (memoized !== undefined) return memoized;
-		}
-		const { SourceProcessor: CssSourceProcessor } =
-			_cssSyntax || (_cssSyntax = require("../css/syntax"));
-		let code;
-		try {
-			code = new CssSourceProcessor().process(
-				source,
-				isBlock ? block : sheet
-			).code;
-		} catch (_err) {
-			// Text the minifier cannot read is left as written, by declining.
-			return undefined;
-		}
-		if (isBlock) blocks.set(source, code);
-		return code;
-	};
-};
 
 /**
  * Extracts `url(...)` references from a CSS value (an SVG presentation
@@ -14000,6 +13918,9 @@ const A = {
 };
 
 module.exports.A = A;
+// The `as` a `style=""` is offered under: a block's contents rather than a
+// whole stylesheet, which is what gives it a rule block's merges.
+module.exports.BLOCK_CONTENTS = BLOCK_CONTENTS;
 // The `as` a classic `<script>` is offered under — the production every
 // JavaScript engine reads by default.
 module.exports.CLASSIC_SCRIPT = CLASSIC_SCRIPT;
@@ -14007,6 +13928,9 @@ module.exports.EMBEDDED_LANGUAGES = EMBEDDED_LANGUAGES;
 // The `as` an event handler attribute is offered under, read by a renderer
 // whose engine takes only whole scripts.
 module.exports.EVENT_HANDLER = EVENT_HANDLER;
+// Neither a JSON `<script>` body nor an `<svg>` subtree is a module source
+// type, so both name themselves here.
+module.exports.JSON_TYPE = JSON_TYPE;
 // The `as` a `<script type=module>` is offered under: the goal symbol decides
 // what parses, so a renderer cannot read it off the body.
 module.exports.MODULE_SCRIPT = MODULE_SCRIPT;
